requirements.txt
tensorflow-gpu>=1.14
tensorflow-hub>=0.4.0
h5py
nltk
pyhocon
scipy
sklearn
